+++ js/mocks.js
@@ -1,0 +1,52 @@
+
+const DESCRIPTION = [
+  'Загадочный взгляд',
+  'Улыбка, которая согревает сердце',
+  'Момент счастья',
+  'Нежность в каждом движении',
+  'Радость встречи',
+  'Волшебство момента',
+  'Искренние эмоции',
+  'Взгляд в будущее',
+  'Минута покоя',
+  'Счастье в мелочах',
+  'Любовь в каждом кадре',
+  'Неотразимая красота',
+  'Вдохновение в каждом снимке',
+  'Забота и теплота',
+  'Путешествие в мир эмоций',
+  'Магия момента',
+  'Радость жизни',
+  'Нежность и забота',
+  'Любовь и гармония',
+  'Забота и поддержка',
+  'Счастье в каждом кадре',
+  'Умиротворение и покой',
+  'Любовь и дружба',
+  'Нежность и искренность',
+  'Волшебство моментальной съёмки'
+];
+
+const NAME = [
+  'Иван',
+  'Мария',
+  'Александр',
+  'Екатерина',
+  'Дмитрий',
+  'Анастасия',
+  'Алексей',
+  'Ольга',
+  'Владимир',
+  'Татьяна'
+];
+
+const COMMENTS = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export { DESCRIPTION, NAME, COMMENTS };
