+++ js/variables.js
@@ -1,0 +1,20 @@
+const PHOTO_AMOUNT = 25;
+const LIKES_AMOUNT = {
+  minLikes: 15,
+  maxLikes: 200
+};
+const AVATAR_ADDRESS_NUMBERS = {
+  min: 1,
+  max: 6
+};
+const COMMENTS_ID_AMOUNT = {
+  min: 0,
+  max: 10000
+};
+
+const COMMENTS_AMOUNT = {
+  min: 0,
+  max: 30
+};
+
+export { PHOTO_AMOUNT, LIKES_AMOUNT, AVATAR_ADDRESS_NUMBERS, COMMENTS_ID_AMOUNT, COMMENTS_AMOUNT };
